+++ src/HoleFillingPackage/WeightingFunction/DefaultWeightingFunctionFactory.java
@@ -3,7 +3,7 @@ package HoleFillingPackage.WeightingFunction;
 public class DefaultWeightingFunctionFactory implements  IWeightingFunctionFactory {
 
     @Override
-    public IWeightingFunction Create(int z, double epsilon) {
+    public IWeightingFunction Create(int z, float epsilon) {
         return new DefaultWeightingFunction(z, epsilon);
     }
 }
